detection2 copy/requirement.txt
Flask
Flask-CORS
torch
transformers
scikit-learn
joblib
mysql-connector-python
beautifulsoup4
numpy
urllib3
python-dotenv